refactor: Remove phishing protection and reorganize component tests
- Removed incomplete phishing protection feature (only blocked 1 domain)
- Deleted phishing detection utilities, components, and tests
- Removed phishing checks from background, provider, and connection services
- Renamed ComposeForm to ComposerForm for better clarity
- Moved all component test files to __tests__ subdirectories
- Fixed test import paths after reorganization

The phishing protection was half-implemented and provided false security.
Better to rely on browser's built-in protection until a proper implementation
with comprehensive domain lists and detection algorithms can be added.
