+++ src/pages/compose/dispenser/close-by-hash/form.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useFormStatus } from "react-dom";
 import { Field, Label, Description, Textarea } from "@headlessui/react";
-import { ComposeForm } from "@/components/compose-form";
+import { ComposerForm } from "@/components/composer-form";
 import { AddressHeader } from "@/components/headers/address-header";
 import { useComposer } from "@/contexts/composer-context";
 import { fetchDispenserByHash } from "@/utils/blockchain/counterparty";
@@ -86,7 +86,7 @@ export function DispenserCloseByHashForm({
   };
 
   return (
-    <ComposeForm
+    <ComposerForm
       formAction={formAction}
       header={
         activeAddress && (
@@ -133,6 +133,6 @@ export function DispenserCloseByHashForm({
             </Field>
 
       )}
-    </ComposeForm>
+    </ComposerForm>
   );
 }
