+++ src/pages/compose/issuance/form.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from "react";
 import { useFormStatus } from "react-dom";
 import { Field, Label, Description, Input, Textarea } from "@headlessui/react";
-import { ComposeForm } from "@/components/compose-form";
+import { ComposerForm } from "@/components/composer-form";
 import { CheckboxInput } from "@/components/inputs/checkbox-input";
 import { AssetNameInput } from "@/components/inputs/asset-name-input";
 import { AmountWithMaxInput } from "@/components/inputs/amount-with-max-input";
@@ -120,7 +120,7 @@ export function IssuanceForm({
   };
 
   return (
-    <ComposeForm
+    <ComposerForm
       formAction={async formData => {
         // If inscribing, convert file to base64 and set as description
         if (inscribeEnabled) {
@@ -281,6 +281,6 @@ export function IssuanceForm({
             />
           )}
 
-    </ComposeForm>
+    </ComposerForm>
   );
 }
