+++ src/pages/compose/issuance/lock-supply/form.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from "react";
 import { useFormStatus } from "react-dom";
-import { ComposeForm } from "@/components/compose-form";
+import { ComposerForm } from "@/components/composer-form";
 import { CheckboxInput } from "@/components/inputs/checkbox-input";
 import { useComposer } from "@/contexts/composer-context";
 import { useAssetInfo } from "@/hooks/useAssetInfo";
@@ -41,7 +41,7 @@ export function LockSupplyForm({
   }
 
   return (
-    <ComposeForm
+    <ComposerForm
       formAction={formAction}
     >
       <div className="mb-4 p-3 bg-gray-50 rounded-md">
@@ -66,6 +66,6 @@ export function LockSupplyForm({
           disabled={pending}
         />
 
-    </ComposeForm>
+    </ComposerForm>
   );
 }
