+++ src/pages/compose/utxo/detach/form.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState, useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import { FaSpinner } from "react-icons/fa";
 import { Field, Label, Description, Input } from "@headlessui/react";
-import { ComposeForm } from "@/components/compose-form";
+import { ComposerForm } from "@/components/composer-form";
 import { AddressHeader } from "@/components/headers/address-header";
 import { DestinationInput } from "@/components/inputs/destination-input";
 import { useComposer } from "@/contexts/composer-context";
@@ -69,7 +69,7 @@ export function UtxoDetachForm({
   }, [initialUtxo, initialFormData?.sourceUtxo]);
 
   return (
-    <ComposeForm
+    <ComposerForm
       formAction={formAction}
       header={
         activeAddress && (
@@ -135,6 +135,6 @@ export function UtxoDetachForm({
             helpText="The address to detach assets to. If not provided, assets will be detached to the UTXO's owner address."
           />
 
-    </ComposeForm>
+    </ComposerForm>
   );
 }
